requirements.txt
django==4.2
djangorestframework
markdown
django-filter
djangorestframework-simplejwt